BRA: block constants are not taped, delays with a non-literal amount, a 64-bit dsp layout
Three limits met by a 16-line differentiable FDN (170 RBJ sections, 279
parameters, one rad over a 48 kHz response), and a forward bug found on the way.

- is_trivially_reverse_evaluable gets Min/Max/Pow/Atan2/Fmod/Remainder/Select2
  of stateless operands: ma.SR is min(192000, max(1, fconstant(...))), so every
  coefficient computed from the sample rate was taped per sample. One tape per
  lowered FIR value (tape_by_value). One RBJ section: 21 tapes -> 5; the FDN:
  4449 -> 1391, 3.2 s -> 1.0 s per 65536-frame epoch.
- Delay(d, x) with a non-literal amount has a reverse rule
  (propagate_bra_variable_delay_adj): the adjoint is scattered to n - d[n]
  through a D + 1 slot buffer, the amount replayed from its tape when it varies
  (collect_delay_amounts). It used to be read as zero, the delay as the
  identity, and the gradients through a recursion holding it were silently
  wrong. A non-literal amount read directly on a recursion output is rejected.
- lower_fixed_delay's ExactShift arm read the recursion's shift array at the
  sizing bound whatever the runtime amount: 0.5 : + ~ (@(int(d)) : *(g))
  stayed at 0.5 for ever. It reads amount + implicit_delay.
- The Cranelift dsp* layout offsets and sizes are u64 (field_offset_imm):
  the tapes of a rad over a long block put the state beyond 4 GB. The FDN
  runs on a 345600-frame block (7.2 s, -bra-tape 524288), 5.8 GB of tapes.

Tests: rad_runtime (slider amount in a loop, time-varying amount in a loop,
fractional amount as a seed, against finite differences), bra_tapes (tape
counts), recursion_delay_amount (the forward against the literal program and
the recurrence), a 4 GiB layout in codegen. Docs: rad-note §4, rad-usage.
Journal: porting/journal/2026-09-10.md.

+++ crates/cranelift-ffi/src/types.rs
@@ -174,7 +174,7 @@ impl DspStateBuffer {
             .iter()
             .find(|zone| zone.role == MemoryRole::DspObject)
             .ok_or_else(|| "mem0 layout has no DSP object zone".to_owned())?;
-        if object.size_bytes != u64::from(layout.size_bytes())
+        if object.size_bytes != layout.size_bytes()
             || object.alignment != u64::from(layout.align_bytes())
         {
             return Err(format!(
